Add tests as per requirements

+++ src/MunroLibrary.Api/Dtos/MunroQueryDto.cs
@@ -9,8 +9,8 @@ namespace MunroLibrary.Api.Dtos
         public int PageNumber { get; set; } = 1;
         public int PageSize { get; set; } = 1000;
         public MunroType? MunroType { get; set; }
-        public int? MinHeight { get; set; }
-        public int? MaxHeight { get; set; }
+        public decimal? MinHeight { get; set; }
+        public decimal? MaxHeight { get; set; }
 
     }
 }
